modularizes interfaces from createsNodesAndEdges to their own files
