+++ src/visivue-browser/src/types/Edge.vue
@@ -1,0 +1,17 @@
+<script lang="ts">
+interface Edge {
+  id: string;
+  label: string;
+  source: string;
+  target: string;
+  style: string;
+  animated: boolean;
+}
+
+interface Style {
+  stroke: string;
+  strokeWidth: string;
+}
+
+export { Edge, Style };
+</script>
